put function for creating asset nodes from data in its own file
