+++ packages/gatsby-transformer-cloudinary/gatsby-node.js
@@ -6,7 +6,6 @@ const {
 const { uploadImageNodeToCloudinary } = require('./upload');
 const { setPluginOptions } = require('./options');
 const { createImageNode } = require('./create-image-node');
-const { createCloudinaryAssetNode } = require('./create-cloudinary-asset-node');
 
 const ALLOWED_MEDIA_TYPES = ['image/png', 'image/jpeg', 'image/gif'];
 
@@ -168,47 +167,6 @@ async function createAssetNodeFromFile({
   });
 
   return imageNode;
-}
-
-function getAssetDataKeys(node) {
-  return Object.keys(node).filter(key => {
-    return (
-      node[key] &&
-      node[key].cloudinaryAssetData === true &&
-      node[key].cloudName &&
-      node[key].publicId &&
-      node[key].originalHeight &&
-      node[key].originalWidth
-    );
-  });
-}
-
-async function createAssetNodesFromData({
-  node,
-  actions: { createNode },
-  createNodeId,
-  createContentDigest,
-  reporter,
-}) {
-  const assetDataKeys = getAssetDataKeys(node);
-
-  assetDataKeys.forEach(assetDataKey => {
-    const assetData = { ...node[assetDataKey] };
-    delete node[assetDataKey];
-    createCloudinaryAssetNode({
-      cloudName: assetData.cloudName,
-      createContentDigest,
-      createNode,
-      createNodeId,
-      originalHeight: assetData.originalHeight,
-      originalWidth: assetData.originalWidth,
-      parentNode: node,
-      publicId: assetData.publicId,
-      relationshipName: assetDataKey,
-      reporter,
-      version: assetData.version,
-    });
-  });
 }
 
 exports.onCreateNode = async ({
